+++ python/fnllm/tests/unit/openai/llm/test_chat_text.py
@@ -7,7 +7,6 @@ from fnllm.openai.llm.openai_text_chat_llm import OpenAITextChatLLMImpl
 def test_child_with_cache():
     llm = OpenAITextChatLLMImpl(
         client=Mock(),
-        cache=None,
         model="model",
         json_receiver=None,
         json_strategy=JsonStrategy.VALID,
@@ -17,7 +16,7 @@ def test_child_with_cache():
 
     llm = OpenAITextChatLLMImpl(
         client=Mock(),
-        cache=Mock(),
+        cached=Mock(),
         model="model",
         json_receiver=None,
         json_strategy=JsonStrategy.LOOSE,
@@ -29,7 +28,6 @@ def test_child_with_cache():
 def test_is_reasoning_model():
     llm = OpenAITextChatLLMImpl(
         client=Mock(),
-        cache=None,
         model="o1-mini",
         json_receiver=None,
         json_strategy=JsonStrategy.VALID,
@@ -42,7 +40,6 @@ def test_is_reasoning_model():
 def test_is_not_reasoning_model():
     llm = OpenAITextChatLLMImpl(
         client=Mock(),
-        cache=None,
         model="model",
         json_receiver=None,
         json_strategy=JsonStrategy.VALID,